[#74] Refactored Tests (#77)
