+++ Tests/YMFFTests/Stores/MutableStoreTests.swift
@@ -13,8 +13,6 @@ import YMFFProtocols
 
 @testable import YMFF
 
-// MARK: - MutableStoreTests
-
 final class MutableStoreTests: XCTestCase {
     
     private var mutableStore: MutableFeatureFlagStoreProtocol!
@@ -24,7 +22,7 @@ final class MutableStoreTests: XCTestCase {
         super.setUp()
         
         mutableStore = MutableFeatureFlagStore(store: .init())
-        resolver = FeatureFlagResolver(configuration: .init(stores: [.mutable(mutableStore)]))
+        resolver = FeatureFlagResolver(stores: [.mutable(mutableStore)])
     }
     
     func testOverride() {
@@ -77,53 +75,6 @@ final class MutableStoreTests: XCTestCase {
         resolver = nil
         
         XCTAssertEqual(saveChangesCount, 1)
-    }
-    
-}
-
-// MARK: - MutableFeatureFlagStore
-
-final private class MutableFeatureFlagStore: MutableFeatureFlagStoreProtocol {
-    
-    private var store: TransparentFeatureFlagStore
-    private var onSaveChangesClosure: () -> Void
-    
-    init(store: TransparentFeatureFlagStore, onSaveChanges: @escaping () -> Void = { }) {
-        self.store = store
-        self.onSaveChangesClosure = onSaveChanges
-    }
-    
-    func containsValue(forKey key: String) -> Bool {
-        store[key] != nil
-    }
-    
-    func value<Value>(forKey key: String) -> Value? {
-        let expectedValueType = Value.self
-        
-        switch expectedValueType {
-        case is Bool.Type,
-             is Int.Type,
-             is String.Type,
-             is Optional<Bool>.Type,
-             is Optional<Int>.Type,
-             is Optional<String>.Type:
-            return store[key] as? Value
-        default:
-            assertionFailure("The expected feature flag value type (\(expectedValueType)) is not supported")
-            return nil
-        }
-    }
-    
-    func setValue<Value>(_ value: Value, forKey key: String) {
-        store[key] = value
-    }
-    
-    func removeValue(forKey key: String) {
-        store.removeValue(forKey: key)
-    }
-    
-    func saveChanges() {
-        onSaveChangesClosure()
     }
     
 }
